api/requirements.txt
requests
python-dotenv
fastapi
pydantic
uvicorn
pymongo
pymongo[srv]
aiofiles
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytz
twilio
boto3
jinja2